backend/requirements.txt
face_recognition==1.3.0
opencv-python==4.11.0.86
dlib==19.24.1
numpy==1.26.2
fastapi==0.95.0 
uvicorn==0.22.0 
python-dotenv==1.0.0 
pymysql==1.1.0 
ngrok==0.2.5 
opencv-python==4.8.0 
face-recognition==1.3.0 
numpy==1.24.0 
sqlalchemy==2.0.0 
